+++ apps/synapse-playground/src/components/warm-storage/uploads-section.tsx
@@ -2,16 +2,16 @@ import { type DataSetWithPieces, type UseProvidersResult, useUpload } from '@fil
 import { useEffect, useState } from 'react'
 import { ErrorAlert, HashAlert } from '../custom-ui/alerts.tsx'
 import { ButtonLoading } from '../custom-ui/button-loading.tsx'
-import { Label } from '../ui/label.tsx'
-import {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectLabel,
-  SelectTrigger,
-  SelectValue,
-} from '../ui/select.tsx'
+// import { Label } from '../ui/label.tsx'
+// import {
+//   Select,
+//   SelectContent,
+//   SelectGroup,
+//   SelectItem,
+//   SelectLabel,
+//   SelectTrigger,
+//   SelectValue,
+// } from '../ui/select.tsx'
 import { Skeleton } from '../ui/skeleton.tsx'
 import { UploadsZone } from './uploads-zone.tsx'
 
@@ -26,10 +26,10 @@ export function UploadsSection({
   const [dataSet, setDataSet] = useState<string | undefined>(undefined)
   const [files, setFiles] = useState<File[] | undefined>()
 
-  const providerWithDataSets = providers?.map((provider) => ({
-    ...provider,
-    dataSets: dataSets?.filter((d) => d.providerId === provider.id),
-  }))
+  // const providerWithDataSets = providers?.map((provider) => ({
+  //   ...provider,
+  //   dataSets: dataSets?.filter((d) => d.providerId === provider.id),
+  // }))
   const approvedProviders = providers?.map((p) => p.id)
 
   const dataSetsFromApprovedProviders = dataSets?.filter((d) => approvedProviders?.includes(d.providerId))
@@ -45,8 +45,11 @@ export function UploadsSection({
     isPending: isUploading,
     error: uploadError,
   } = useUpload({
-    onHash: (hash) => {
-      setHash(hash)
+    source: 'synapse-playground',
+    callbacks: {
+      onPiecesAdded(transaction) {
+        setHash(transaction)
+      },
     },
     mutation: {
       onSettled: () => {
@@ -60,7 +63,7 @@ export function UploadsSection({
     setFiles(files)
   }
 
-  function onSubmit(e: React.FormEvent<HTMLFormElement>) {
+  function onSubmit(e: React.SubmitEvent<HTMLFormElement>) {
     e.preventDefault()
 
     if (files && files.length > 0 && dataSet) {
@@ -69,14 +72,21 @@ export function UploadsSection({
       //   dataSetId,
       //   sessionKey: sessionKey,
       // })
-      upload({ files: Array.from(files), dataSetId: BigInt(dataSet) })
+      const file = files[0]
+      upload({
+        file,
+        metadata: {
+          name: file.name,
+          type: file.type,
+        },
+      })
     }
   }
 
   return dataSet ? (
     <form onSubmit={onSubmit}>
       <div className="flex flex-col gap-3 my-4">
-        <Label htmlFor="data-set">Data Set</Label>
+        {/* <Label htmlFor="data-set">Data Set</Label>
         <Select
           name="data-set"
           onValueChange={(value) => {
@@ -99,7 +109,7 @@ export function UploadsSection({
               </SelectGroup>
             ))}
           </SelectContent>
-        </Select>
+        </Select> */}
         <UploadsZone files={files} handleDrop={handleDrop} />
         <ButtonLoading disabled={!dataSet} loading={isUploading} type="submit">
           Upload